Move DDIR to its own crate
Extract the DD IR interpreter into a `ddir/` crate, separating shared
infrastructure (parse, IR, lower) from backend-specific rendering.

- Split parse into `parse/applicative.rs` (.ddir) and `parse/pipe.rs` (.ddp)
- Extract `ir.rs` (Node, Program, RowLike) from `lower.rs`
- Move programs and examples from differential-dataflow/examples/ to ddir/
- Remove DDIR_DEBUG, DDIR_REACHABILITY, DDIR_PRINT env vars (use INSPECT)
- Fix off-by-one in columnar harness timestamping
- Unify inspect output format between vec and col backends
- Update programs to output compact counts via arrange+inspect
