+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
     #"doop",
     "mdbook",
     "diagnostics",
+    "ddir",
 ]
 resolver = "2"
