+++ ddir/examples/programs/kcore.ddir
