+++ ddir/examples/programs/scc.ddp
